Allow creating `Topic`s in the same way as `Target`s; reorganize `Targets` utility class into a package-level class instead of nested one.
